+++ clients/mobile/components/tasks/assignee-picker.tsx
@@ -1,11 +1,13 @@
-import { useState } from "react";
-import { Text } from "react-native";
-import { UserRound } from "lucide-react-native";
+import { useRef, useState } from "react";
+import { View, Text, Pressable, TextInput } from "react-native";
+import { UserRound, ChevronRight, Search, X } from "lucide-react-native";
 import { useQuery } from "@tanstack/react-query";
 import { useAPIClient } from "@shared/api/client";
 import { getConfig } from "@shared/api/config";
 import type { User } from "@shared";
-import { SearchPicker } from "@/components/tasks/search-picker";
+import { Colors } from "@/constants/theme";
+
+const ICON_COLOR = Colors.light.textSubtle;
 
 type SearchUsersResponse = {
   users: Array<User>;
@@ -18,8 +20,9 @@ type AssigneePickerProps = {
 };
 
 export function AssigneePicker({ value, onChange }: AssigneePickerProps) {
+  const [expanded, setExpanded] = useState(false);
   const [search, setSearch] = useState("");
-  const [queryEnabled, setQueryEnabled] = useState(false);
+  const inputRef = useRef<TextInput>(null);
 
   const api = useAPIClient();
   const { hotelId } = getConfig();
@@ -31,43 +34,113 @@ export function AssigneePicker({ value, onChange }: AssigneePickerProps) {
         hotel_id: hotelId,
         q: search || undefined,
       }),
-    enabled: queryEnabled,
+    enabled: expanded,
   });
 
   const users = data?.users ?? [];
 
-  function getDisplayName(user: User) {
-    return [user.first_name, user.last_name].filter(Boolean).join(" ");
+  function handleToggle() {
+    const next = !expanded;
+    setExpanded(next);
+    setSearch("");
+    if (next) setTimeout(() => inputRef.current?.focus(), 50);
   }
 
+  function handleSelect(user: User) {
+    onChange(value?.id === user.id ? undefined : user);
+    setExpanded(false);
+    setSearch("");
+  }
+
+  const displayName = value
+    ? [value.first_name, value.last_name].filter(Boolean).join(" ")
+    : undefined;
+
   return (
-    <SearchPicker
-      icon={UserRound}
-      label="Assignee"
-      value={value}
-      onChange={onChange}
-      getKey={(u) => u.id ?? ""}
-      getTriggerLabel={getDisplayName}
-      getChipLabel={getDisplayName}
-      renderRow={(user) => (
-        <>
-          <Text className="text-xs text-text-default tracking-tight">
-            {getDisplayName(user)}
+    <View className="gap-2">
+      <Pressable
+        onPress={handleToggle}
+        className="flex-row items-center justify-between h-6"
+      >
+        <View className="flex-row items-center gap-1">
+          <UserRound size={16} color={ICON_COLOR} />
+          <Text className="text-[15px] text-text-subtle tracking-tight">
+            Assignee
           </Text>
-          {user.role && (
-            <Text className="text-xs text-text-subtle tracking-tight">
-              {user.role}
-            </Text>
+        </View>
+        <View className="flex-row items-center gap-1">
+          <Text
+            className={`text-[15px] tracking-tight ${displayName ? "text-text-default" : "text-text-subtle"}`}
+          >
+            {displayName ?? "Select..."}
+          </Text>
+          <ChevronRight
+            size={14}
+            color={ICON_COLOR}
+            style={{ transform: [{ rotate: expanded ? "90deg" : "0deg" }] }}
+          />
+        </View>
+      </Pressable>
+
+      {expanded && (
+        <View className="rounded border border-input-border">
+          <View className="flex-row items-center gap-2.5 px-2 py-2 border-b border-input-border-light">
+            <Search size={17} color={ICON_COLOR} />
+            <View className="flex-row items-center flex-1 gap-1">
+              {value && (
+                <View className="flex-row items-center gap-1 bg-bg-selected rounded px-1.5 py-0.5">
+                  <Text className="text-xs text-text-default tracking-tight">
+                    {displayName}
+                  </Text>
+                  <Pressable onPress={() => onChange(undefined)} hitSlop={4}>
+                    <X size={10} color={ICON_COLOR} />
+                  </Pressable>
+                </View>
+              )}
+              <TextInput
+                ref={inputRef}
+                className="flex-1 text-xs text-text-default tracking-tight"
+                placeholder="Search by name..."
+                placeholderTextColor={ICON_COLOR}
+                value={search}
+                onChangeText={setSearch}
+                returnKeyType="search"
+              />
+            </View>
+          </View>
+
+          {isLoading ? (
+            <View className="px-4 py-3">
+              <Text className="text-xs text-text-subtle tracking-tight">
+                Loading...
+              </Text>
+            </View>
+          ) : users.length === 0 ? (
+            <View className="px-4 py-3">
+              <Text className="text-xs text-text-subtle tracking-tight">
+                No staff found
+              </Text>
+            </View>
+          ) : (
+            users.map((user) => (
+              <Pressable
+                key={user.id}
+                onPress={() => handleSelect(user)}
+                className={`px-4 py-2 ${value?.id === user.id ? "bg-bg-selected" : ""}`}
+              >
+                <Text className="text-xs text-text-default tracking-tight">
+                  {[user.first_name, user.last_name].filter(Boolean).join(" ")}
+                </Text>
+                {user.role && (
+                  <Text className="text-xs text-text-subtle tracking-tight">
+                    {user.role}
+                  </Text>
+                )}
+              </Pressable>
+            ))
           )}
-        </>
+        </View>
       )}
-      items={users}
-      isLoading={isLoading}
-      search={search}
-      onSearch={setSearch}
-      onExpandChange={setQueryEnabled}
-      searchPlaceholder="Search by name..."
-      emptyMessage="No staff found"
-    />
+    </View>
   );
 }
